Move PANTS_WORKDIR_SUFFIX to build_environment
