+++ src/python/pants/option/global_options.py
@@ -8,15 +8,12 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import logging
 import os
 
-from pants.base.build_environment import (get_buildroot, get_pants_cachedir, get_pants_configdir,
-                                          pants_version)
+from pants.base.build_environment import (PANTS_WORKDIR_SUFFIX, get_buildroot, get_pants_cachedir,
+                                          get_pants_configdir, pants_version)
 from pants.option.arg_splitter import GLOBAL_SCOPE
 from pants.option.custom_types import list_option
 from pants.option.optionable import Optionable
 from pants.option.scope import ScopeInfo
-
-
-PANTS_WORKDIR_SUFFIX = '.pants.d'
 
 
 class GlobalOptionsRegistrar(Optionable):
